+++ backends/bevy_picking_rapier/src/lib.rs
@@ -119,12 +119,13 @@ pub fn update_hits(
                     QueryFilter::new().predicate(&|entity| {
                         let marker_requirement =
                             !backend_settings.require_markers || marked_targets.get(entity).is_ok();
-                        let render_layers_match = match (cam_layers, layers.get(entity)) {
-                            (Some(cam_layers), Ok(entity_layers)) => {
-                                cam_layers.intersects(entity_layers)
-                            }
-                            _ => true, // If either `RenderLayers` components is not present, ignore.
-                        };
+
+                        // Cameras missing render layers intersect all layers
+                        let cam_layers = cam_layers.copied().unwrap_or(RenderLayers::all());
+                        // Other entities missing render layers are on the default layer 0
+                        let entity_layers = layers.get(entity).copied().unwrap_or_default();
+                        let render_layers_match = cam_layers.intersects(&entity_layers);
+
                         let pickable = pickables
                             .get(entity)
                             .map(|p| *p != Pickable::IGNORE)